Move cuda/runtime/shim to cuda/runtime for better usibility.
This diff moves the `cuda/runtime/shim` directory to `cuda/runtime` to make the utility functions sharable among all runtime infrastructure.

Differential Revision: [D84122729](https://our.internmc.facebook.com/intern/diff/D84122729/)

ghstack-source-id: 314747331
Pull Request resolved: https://github.com/pytorch/executorch/pull/14879
