+++ backends/cuda/runtime/TARGETS
@@ -11,7 +11,7 @@ runtime.cxx_library(
     headers = [
         "shims/memory.h",
         "shims/tensor_attribute.h",
-        "shims/utils.h",
+        "utils.h",
     ],
     # @lint-ignore BUCKLINT: Avoid `link_whole=True` (https://fburl.com/avoid-link-whole)
     link_whole = True,
